backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "pillow>=12.0.0",
]
